pyproject.toml
[project]
name = "autoddg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.8.*"
dependencies = [
    "colorama>=0.4.6",
    "datumaro>=1.5.2",
    "numpy>=1.24.4",
    "opencv-python>=4.10.0.84",
    "pandas>=2.0.3",
    "pycocotools>=2.0.7",
    "pyqt5>=5.15.11",
    "pyqt6>=6.7.1",
    "python-dotenv>=1.0.1",
    "randomname>=0.2.1",
    "rasterio>=1.3.11",
    "ruff>=0.9.2",
    "scikit-image>=0.21.0",
    "scikit-learn>=1.3.2",
    "torch>=2.4.1",
    "torchaudio>=2.4.1",
    "torchvision>=0.19.1",
    "tqdm>=4.67.1",
]
